+++ app/src/main/java/com/github/se/studybuddies/database/ServiceLocator.kt
@@ -4,15 +4,24 @@ import com.google.firebase.auth.FirebaseAuth
 
 object ServiceLocator {
   private var dbRepository: DbRepository? = null
+  private var realtimeRepository: RealtimeRepository? = null
   private var currentUserUID: String? = null
 
   fun provideDatabase(): DbRepository {
     return dbRepository ?: DatabaseConnection()
   }
 
+    fun provideRealtimeDatabase(): RealtimeRepository {
+        return realtimeRepository ?: RealtimeConnection()
+    }
+
   fun setMockDatabase(dbRepository: DbRepository) {
     this.dbRepository = dbRepository
   }
+
+    fun setMockRealtimeDatabase(realtimeRepository: RealtimeRepository) {
+        this.realtimeRepository = realtimeRepository
+    }
 
   fun setCurrentUserUID(uid: String) {
     currentUserUID = uid
