+++ app/src/main/java/com/github/se/studybuddies/viewModels/DirectMessageViewModel.kt
@@ -3,7 +3,7 @@ package com.github.se.studybuddies.viewModels
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.se.studybuddies.data.Chat
-import com.github.se.studybuddies.database.DbRepository
+import com.github.se.studybuddies.database.RealtimeRepository
 import com.github.se.studybuddies.database.ServiceLocator
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -12,8 +12,8 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 
 class DirectMessageViewModel(
-    private val userUid: String = "",
-    private val db: DbRepository = ServiceLocator.provideDatabase()
+    userUid: String = "",
+    private val db: RealtimeRepository = ServiceLocator.provideRealtimeDatabase()
 ) : ViewModel() {
   private val _userUid = MutableStateFlow(userUid)
   private val _directMessages = MutableStateFlow<List<Chat>>(emptyList())
